move DisplaySettings.cs
