+++ src/game/Settings/JumpvalleySettings.cs
@@ -1,3 +1,5 @@
+using JumpvalleyGame.Settings.Display;
+
 namespace JumpvalleyGame.Settings
 {
     /// <summary>
